feat: monitor guardian health
