+++ fmo_server/src/federation/mod.rs
@@ -1,4 +1,5 @@
 pub mod db;
+mod guardians;
 mod meta;
 pub mod observer;
 mod session;
